+++ air/src/constraints/op_flags/mod.rs
@@ -21,7 +21,13 @@
 
 use core::marker::PhantomData;
 
-use miden_core::{field::PrimeCharacteristicRing, operations::Operation};
+use miden_core::{
+    field::PrimeCharacteristicRing,
+    operations::{
+        OPCODE_CALL, OPCODE_DYN, OPCODE_DYNCALL, OPCODE_END, OPCODE_HALT, OPCODE_JOIN, OPCODE_LOOP,
+        OPCODE_REPEAT, OPCODE_RESPAN, OPCODE_SPAN, OPCODE_SPLIT, OPCODE_SYSCALL, Operation,
+    },
+};
 
 #[cfg(test)]
 use crate::trace::decoder::NUM_OP_BITS;
@@ -230,7 +236,7 @@ macro_rules! op_flag_getters {
             $(#[$meta])*
             #[inline(always)]
             pub fn $name(&self) -> E {
-                self.$array[get_op_index($op.op_code())].clone()
+                self.$array[get_op_index($op)].clone()
             }
         )*
     };
@@ -635,208 +641,208 @@ where
     op_flag_getters!(degree7_op_flags,
         /// Operation Flag of NOOP operation.
         #[allow(dead_code)]
-        noop => Operation::Noop,
+        noop => Operation::Noop.op_code(),
         /// Operation Flag of EQZ operation.
-        eqz => Operation::Eqz,
+        eqz => Operation::Eqz.op_code(),
         /// Operation Flag of NEG operation.
-        neg => Operation::Neg,
+        neg => Operation::Neg.op_code(),
         /// Operation Flag of INV operation.
-        inv => Operation::Inv,
+        inv => Operation::Inv.op_code(),
         /// Operation Flag of INCR operation.
-        incr => Operation::Incr,
+        incr => Operation::Incr.op_code(),
         /// Operation Flag of NOT operation.
-        not => Operation::Not,
+        not => Operation::Not.op_code(),
         /// Operation Flag of MLOAD operation.
-        mload => Operation::MLoad,
+        mload => Operation::MLoad.op_code(),
         /// Operation Flag of SWAP operation.
-        swap => Operation::Swap,
+        swap => Operation::Swap.op_code(),
         /// Operation Flag of CALLER operation.
         ///
         /// CALLER overwrites the top 4 stack elements with the hash of the function
         /// that initiated the current SYSCALL.
-        caller => Operation::Caller,
+        caller => Operation::Caller.op_code(),
         /// Operation Flag of MOVUP2 operation.
-        movup2 => Operation::MovUp2,
+        movup2 => Operation::MovUp2.op_code(),
         /// Operation Flag of MOVDN2 operation.
-        movdn2 => Operation::MovDn2,
+        movdn2 => Operation::MovDn2.op_code(),
         /// Operation Flag of MOVUP3 operation.
-        movup3 => Operation::MovUp3,
+        movup3 => Operation::MovUp3.op_code(),
         /// Operation Flag of MOVDN3 operation.
-        movdn3 => Operation::MovDn3,
+        movdn3 => Operation::MovDn3.op_code(),
         /// Operation Flag of ADVPOPW operation.
         #[allow(dead_code)]
-        advpopw => Operation::AdvPopW,
+        advpopw => Operation::AdvPopW.op_code(),
         /// Operation Flag of EXPACC operation.
-        expacc => Operation::Expacc,
+        expacc => Operation::Expacc.op_code(),
         /// Operation Flag of MOVUP4 operation.
-        movup4 => Operation::MovUp4,
+        movup4 => Operation::MovUp4.op_code(),
         /// Operation Flag of MOVDN4 operation.
-        movdn4 => Operation::MovDn4,
+        movdn4 => Operation::MovDn4.op_code(),
         /// Operation Flag of MOVUP5 operation.
-        movup5 => Operation::MovUp5,
+        movup5 => Operation::MovUp5.op_code(),
         /// Operation Flag of MOVDN5 operation.
-        movdn5 => Operation::MovDn5,
+        movdn5 => Operation::MovDn5.op_code(),
         /// Operation Flag of MOVUP6 operation.
-        movup6 => Operation::MovUp6,
+        movup6 => Operation::MovUp6.op_code(),
         /// Operation Flag of MOVDN6 operation.
-        movdn6 => Operation::MovDn6,
+        movdn6 => Operation::MovDn6.op_code(),
         /// Operation Flag of MOVUP7 operation.
-        movup7 => Operation::MovUp7,
+        movup7 => Operation::MovUp7.op_code(),
         /// Operation Flag of MOVDN7 operation.
-        movdn7 => Operation::MovDn7,
+        movdn7 => Operation::MovDn7.op_code(),
         /// Operation Flag of SWAPW operation.
-        swapw => Operation::SwapW,
+        swapw => Operation::SwapW.op_code(),
         /// Operation Flag of MOVUP8 operation.
-        movup8 => Operation::MovUp8,
+        movup8 => Operation::MovUp8.op_code(),
         /// Operation Flag of MOVDN8 operation.
-        movdn8 => Operation::MovDn8,
+        movdn8 => Operation::MovDn8.op_code(),
         /// Operation Flag of SWAPW2 operation.
-        swapw2 => Operation::SwapW2,
+        swapw2 => Operation::SwapW2.op_code(),
         /// Operation Flag of SWAPW3 operation.
-        swapw3 => Operation::SwapW3,
+        swapw3 => Operation::SwapW3.op_code(),
         /// Operation Flag of SWAPDW operation.
-        swapdw => Operation::SwapDW,
+        swapdw => Operation::SwapDW.op_code(),
         /// Operation Flag of EXT2MUL operation.
-        ext2mul => Operation::Ext2Mul,
+        ext2mul => Operation::Ext2Mul.op_code(),
         /// Operation Flag of ASSERT operation.
-        assert_op => Operation::Assert(miden_core::ZERO),
+        assert_op => Operation::Assert(miden_core::ZERO).op_code(),
         /// Operation Flag of EQ operation.
-        eq => Operation::Eq,
+        eq => Operation::Eq.op_code(),
         /// Operation Flag of ADD operation.
-        add => Operation::Add,
+        add => Operation::Add.op_code(),
         /// Operation Flag of MUL operation.
-        mul => Operation::Mul,
+        mul => Operation::Mul.op_code(),
         /// Operation Flag of AND operation.
-        and => Operation::And,
+        and => Operation::And.op_code(),
         /// Operation Flag of OR operation.
-        or => Operation::Or,
+        or => Operation::Or.op_code(),
         /// Operation Flag of U32AND operation.
-        u32and => Operation::U32and,
+        u32and => Operation::U32and.op_code(),
         /// Operation Flag of U32XOR operation.
-        u32xor => Operation::U32xor,
+        u32xor => Operation::U32xor.op_code(),
         /// Operation Flag of DROP operation.
         #[allow(dead_code)]
-        drop => Operation::Drop,
+        drop => Operation::Drop.op_code(),
         /// Operation Flag of CSWAP operation.
-        cswap => Operation::CSwap,
+        cswap => Operation::CSwap.op_code(),
         /// Operation Flag of CSWAPW operation.
-        cswapw => Operation::CSwapW,
+        cswapw => Operation::CSwapW.op_code(),
         /// Operation Flag of MLOADW operation.
-        mloadw => Operation::MLoadW,
+        mloadw => Operation::MLoadW.op_code(),
         /// Operation Flag of MSTORE operation.
-        mstore => Operation::MStore,
+        mstore => Operation::MStore.op_code(),
         /// Operation Flag of MSTOREW operation.
-        mstorew => Operation::MStoreW,
+        mstorew => Operation::MStoreW.op_code(),
         /// Operation Flag of PAD operation.
-        pad => Operation::Pad,
+        pad => Operation::Pad.op_code(),
         /// Operation Flag of DUP operation.
-        dup => Operation::Dup0,
+        dup => Operation::Dup0.op_code(),
         /// Operation Flag of DUP1 operation.
-        dup1 => Operation::Dup1,
+        dup1 => Operation::Dup1.op_code(),
         /// Operation Flag of DUP2 operation.
-        dup2 => Operation::Dup2,
+        dup2 => Operation::Dup2.op_code(),
         /// Operation Flag of DUP3 operation.
-        dup3 => Operation::Dup3,
+        dup3 => Operation::Dup3.op_code(),
         /// Operation Flag of DUP4 operation.
-        dup4 => Operation::Dup4,
+        dup4 => Operation::Dup4.op_code(),
         /// Operation Flag of DUP5 operation.
-        dup5 => Operation::Dup5,
+        dup5 => Operation::Dup5.op_code(),
         /// Operation Flag of DUP6 operation.
-        dup6 => Operation::Dup6,
+        dup6 => Operation::Dup6.op_code(),
         /// Operation Flag of DUP7 operation.
-        dup7 => Operation::Dup7,
+        dup7 => Operation::Dup7.op_code(),
         /// Operation Flag of DUP9 operation.
-        dup9 => Operation::Dup9,
+        dup9 => Operation::Dup9.op_code(),
         /// Operation Flag of DUP11 operation.
-        dup11 => Operation::Dup11,
+        dup11 => Operation::Dup11.op_code(),
         /// Operation Flag of DUP13 operation.
-        dup13 => Operation::Dup13,
+        dup13 => Operation::Dup13.op_code(),
         /// Operation Flag of DUP15 operation.
-        dup15 => Operation::Dup15,
+        dup15 => Operation::Dup15.op_code(),
         /// Operation Flag of ADVPOP operation.
         #[allow(dead_code)]
-        advpop => Operation::AdvPop,
+        advpop => Operation::AdvPop.op_code(),
         /// Operation Flag of SDEPTH operation.
-        sdepth => Operation::SDepth,
+        sdepth => Operation::SDepth.op_code(),
         /// Operation Flag of CLK operation.
-        clk => Operation::Clk,
+        clk => Operation::Clk.op_code(),
     );
 
     // ------ Degree 6 u32 operations  ------------------------------------------------------------
 
     op_flag_getters!(degree6_op_flags,
         /// Operation Flag of U32ADD operation.
-        u32add => Operation::U32add,
+        u32add => Operation::U32add.op_code(),
         /// Operation Flag of U32SUB operation.
-        u32sub => Operation::U32sub,
+        u32sub => Operation::U32sub.op_code(),
         /// Operation Flag of U32MUL operation.
-        u32mul => Operation::U32mul,
+        u32mul => Operation::U32mul.op_code(),
         /// Operation Flag of U32DIV operation.
-        u32div => Operation::U32div,
+        u32div => Operation::U32div.op_code(),
         /// Operation Flag of U32SPLIT operation.
-        u32split => Operation::U32split,
+        u32split => Operation::U32split.op_code(),
         /// Operation Flag of U32ASSERT2 operation.
-        u32assert2 => Operation::U32assert2(miden_core::ZERO),
+        u32assert2 => Operation::U32assert2(miden_core::ZERO).op_code(),
         /// Operation Flag of U32ADD3 operation.
-        u32add3 => Operation::U32add3,
+        u32add3 => Operation::U32add3.op_code(),
         /// Operation Flag of U32MADD operation.
-        u32madd => Operation::U32madd,
+        u32madd => Operation::U32madd.op_code(),
     );
 
     // ------ Degree 5 operations  ----------------------------------------------------------------
 
     op_flag_getters!(degree5_op_flags,
         /// Operation Flag of HPERM operation.
-        hperm => Operation::HPerm,
+        hperm => Operation::HPerm.op_code(),
         /// Operation Flag of MPVERIFY operation.
-        mpverify => Operation::MpVerify(miden_core::ZERO),
+        mpverify => Operation::MpVerify(miden_core::ZERO).op_code(),
         /// Operation Flag of SPLIT operation.
-        split => Operation::Split,
+        split => OPCODE_SPLIT,
         /// Operation Flag of LOOP operation.
-        loop_op => Operation::Loop,
+        loop_op => OPCODE_LOOP,
         /// Operation Flag of SPAN operation.
-        span => Operation::Span,
+        span => OPCODE_SPAN,
         /// Operation Flag of JOIN operation.
-        join => Operation::Join,
+        join => OPCODE_JOIN,
         /// Operation Flag of PUSH operation.
-        push => Operation::Push(miden_core::ONE),
+        push => Operation::Push(miden_core::ONE).op_code(),
         /// Operation Flag of DYN operation.
-        dyn_op => Operation::Dyn,
+        dyn_op => OPCODE_DYN,
         /// Operation Flag of DYNCALL operation.
-        dyncall => Operation::Dyncall,
+        dyncall => OPCODE_DYNCALL,
         /// Operation Flag of EVALCIRCUIT operation.
-        evalcircuit => Operation::EvalCircuit,
+        evalcircuit => Operation::EvalCircuit.op_code(),
         /// Operation Flag of LOG_PRECOMPILE operation.
-        log_precompile => Operation::LogPrecompile,
+        log_precompile => Operation::LogPrecompile.op_code(),
         /// Operation Flag of HORNERBASE operation.
-        hornerbase => Operation::HornerBase,
+        hornerbase => Operation::HornerBase.op_code(),
         /// Operation Flag of HORNEREXT operation.
-        hornerext => Operation::HornerExt,
+        hornerext => Operation::HornerExt.op_code(),
         /// Operation Flag of MSTREAM operation.
-        mstream => Operation::MStream,
+        mstream => Operation::MStream.op_code(),
         /// Operation Flag of PIPE operation.
-        pipe => Operation::Pipe,
+        pipe => Operation::Pipe.op_code(),
     );
 
     // ------ Degree 4 operations  ----------------------------------------------------------------
 
     op_flag_getters!(degree4_op_flags,
         /// Operation Flag of MRUPDATE operation.
-        mrupdate => Operation::MrUpdate,
+        mrupdate => Operation::MrUpdate.op_code(),
         /// Operation Flag of CALL operation.
-        call => Operation::Call,
+        call => OPCODE_CALL,
         /// Operation Flag of SYSCALL operation.
-        syscall => Operation::SysCall,
+        syscall => OPCODE_SYSCALL,
         /// Operation Flag of END operation.
-        end => Operation::End,
+        end => OPCODE_END,
         /// Operation Flag of REPEAT operation.
-        repeat => Operation::Repeat,
+        repeat => OPCODE_REPEAT,
         /// Operation Flag of RESPAN operation.
-        respan => Operation::Respan,
+        respan => OPCODE_RESPAN,
         /// Operation Flag of HALT operation.
-        halt => Operation::Halt,
+        halt => OPCODE_HALT,
         /// Operation Flag of CRYPTOSTREAM operation.
-        cryptostream => Operation::CryptoStream,
+        cryptostream => Operation::CryptoStream.op_code(),
     );
 
     // ------ Composite Flags ---------------------------------------------------------------------